chore: Update Jest config to map `next` export to source correctly
Tests that used the `@metamask/base-controller/next` export were not
working correctly until after a build was completed. They were using
the built code rather than the source code. On CI the tests run without
any build, so they would always fail.

This was resolved up updating the TSConfig and Jest config to map the
`next` export to the correct source code.

This unblocks #6335 as well as other controller migrations to the
upcoming BaseController version.