Speed up network client tests
There are a large amount of tests that exist for the network client that
NetworkController creates and exposes. This is because the network
client assembles a JsonRpcEngine middleware stack, one for Infura RPC
endpoints and one for custom RPC endpoints, and the resulting behavior
must be exercised somehow. Furthermore, we support not only RPC methods
that are in the official Ethereum spec but also ones that fall outside
of the spec, and we must exercise all of those methods too.

Unfortunately, running all of these tests takes a long time in
comparison to others, around 4+ minutes. This is rather painful when
working with `network-controller` as the feedback loop is quite long. We
can reduce this number by splitting up the suite of network client tests
into multiple files. This means that Jest is able to now run them in
parallel.