chore(network-controller): Simplify network switch detection in lookupNetwork
`lookupNetwork` operates on an arbitrary network or, if given nothing,
the globally selected network. The behavior between these two cases is
largely the same, except that in the "selected network" case, it
publishes one of two events, `NetworkController:infuraIsBlocked` or
`NetworkController:infuraIsUnblocked`. This happens after making a
request to the network, and while the request is running, it is possible
for the network to change. This creates a potential problem because if
that happens, the events that we want to fire would no longer represent
the network they were intended to represent.

`lookupNetwork` accounts for this by subscribing to
`NetworkController:networkDidChange` before starting the request; if
it's fired, then it exits early (with the presumption that the network
switch called its own version of `lookupNetwork` and will publish
`NetworkController:infuraIsBlocked` or
`NetworkController:infuraIsUnblocked` at that point in time).

However, this approach to detect a network switch is complicated and
unnecessary. We can simply capture `selectedNetworkClientId` before we
start the request, ask what it is afterward, and compare the two to know
whether the network switched. This means that we can bring the
implementation of the two branches within `lookupNetwork` closer
together (and we can eliminate some tests).