chore(network-controller): Deprecate lookupNetworkByClientId (#6308)
When initializing the globally selected network or switching it to a new
network, we must verify the availability status of the network as well
as its EIP-1559 compatibility. There are two ways to do this in
NetworkController: `lookupNetwork` and `lookupNetworkByClientId`. There
are some slight differences in logic with `lookupNetwork` in how it
handles changes to the globally selected network, but they essentially
do the same thing, and it would make for a less confusing and more
maintainable API if we consolidated them into one method.

Since doing so would be a breaking change, to ease the transition, this
commit deprecates `lookupNetworkByClientId` and suggests that developers
use `lookupNetwork` instead. It also refactors both methods to remove
the duplicate code and ensures that `lookupNetwork` is well-tested both
when given a network client ID and not given one.