Improve sample-controllers to be more useful examples (#6168)
The controllers and services in `sample-controllers` are not as useful
as examples as they could be. They do not follow the best practices that
we want others to follow, so we cannot ask teams to copy them, and they
do not follow patterns that are standard in the product. This commit fixes
that.