Fix API docs generation (#264)
The addition of `web3` as a development dependency in
8aaca96c2b9ddf0efbc826c43d16d5dec0a2d0dd broke `yarn build:docs`, which
generates API docs via TypeDoc and is run when a new version of the
package is published. Curiously, `yarn build` continues to work.

It seems that TypeDoc is using the development variant of the TypeScript
configuration file instead of the build variant, which `yarn build` uses
internally. The build variant limits the scope of TypeScript to just
files within `src`, hiding dependencies. Correcting the TypeDoc config
to use this file fixes the issue.