fix: Keep delegated handlers when unregistering actions (#6395)
## Explanation

Keep delegated handlers around even when unregistering actions. This
helps with the resiliency of the messenger system in case actions are
unregistered and registered elsewhere in the chain of delegations. This
is also very useful for testing.

Additionally adjusts the error message thrown when a delegated action
handler cannot be found to match an undelegated one.