chore(network-controller): Speed up network client tests (#6377)
There are a large amount of tests that exist for the network client in
`network-controller`. This is because the network client assembles a
JsonRpcEngine middleware stack, one for Infura RPC endpoints and one for
custom RPC endpoints, and we must verify the resulting behavior across
not only official Ethereum RPC methods but also MetaMask-specific RPC
methods.

Unfortunately, running all of these tests takes a long time in
comparison to others, around 3.25 minutes. This is rather painful when
working with `network-controller`, as the feedback loop is quite long.
We can reduce this number by splitting up the suite of network client
tests into multiple files. This means that Jest is able to now run them
in parallel. As a result, they now run in about 1.5 minutes, a 2x
speedup.