fix: use signature as txhistory key for Solana bridge transactions (#6424)
## Explanation

This change fixes a bug in Solana post-submission logic in which the
txHistory key is set to a random uuid instead of the tx hash. This
causes txHistory lookups by txHash (for Solana transactions) to fail on
the clients

This is not a breaking change, but requires a migration on the clients
to set the txHistory key to the `historyItem.status?.srcChain?.txHash`
value for Solana transactions. This fixes txHistory items created prior
to this version of the bridge-status-controller

Migration script for extension:
https://github.com/MetaMask/metamask-extension/pull/35539/files#diff-3cd53aba68cf73714b1dbe70345a0b896589f13d8c3cc370d04393a5c6b4f0d8

This PR also fixes properties for the swap Completed events so that they
read data from the txHistory, if it exists

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->

## References

<!--
Are there any issues that this pull request is tied to?
Are there other links that reviewers should consult to understand these
changes better?
Are there client or consumer pull requests to adopt any breaking
changes?

For example:

* Fixes #12345
* Related to #67890
-->

Fixes https://consensyssoftware.atlassian.net/browse/SWAPS-2595

## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've communicated my changes to consumers by [updating changelogs
for packages I've
changed](https://github.com/MetaMask/core/tree/main/docs/contributing.md#updating-changelogs),
highlighting breaking changes as necessary
- [ ] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes