Fix network-controller build
In c4de052f66a53fee451e3e9ca3a648b0123d0be0, some of the
network-controller tests were reorganized. Unfortunately this causes an
issue when using the compiled version of the package in clients because
it seems that we have both a `create-network-client.ts` file and a
`create-nework-client/` directly, and this seems to confuse either
`ts-bridge` or the TypeScript compiler, as it generates a version of
`create-auto-managed-network-client.ts` with an invalid import path for
`create-network-client.ts`. Renaming the `create-nework-client/`
directory seems to fix the issue.