Fix network-controller build (#6452)
In c4de052f66a53fee451e3e9ca3a648b0123d0be0, some of the
`network-controller` tests were reorganized. Unfortunately this causes
an issue when using the compiled version of the package in clients.
Having both a `create-network-client.ts` file and a
`create-nework-client/` directory (with no link between the two) seems
to confuse either `ts-bridge` or the TypeScript compiler, as it
generates a version of `create-auto-managed-network-client.ts` with an
invalid import path for `create-network-client.ts` (that is, it should
end with either `.cjs` or `.mjs` depending on the variant being
generated, but it does not). Renaming the `create-network-client/`
directory seems to fix the issue.