Alert users about incomplete release PRs
Currently we use the `is-release` action to determine whether a commit
merged to `main` is a release commit, and if so, we run the release
workflow. A commit becomes a release commit when:

- the root package version is bumped
- the subject of the merged commit matches "Release X.Y.Z",
  "Release/X.Y.Z", or something similar

However, oftentimes engineers will create a PR and bump some packages'
versions but forget to bump the root version. When this PR is merged,
the release workflow won't run — but by then it is too late, and the
engineer is forced to revert the PR and recreate it correctly. This is
painful.

To prevent this, this commit replaces the `is-action` with a custom
workflow that looks at three things:

- whether the root package is bumped
- whether the title of the release is well-formed
- whether any of the packages in the monorepo have been bumped

As before, this workflow runs when a release commit is merged, and the
result is used to determine whether to run the release workflow. But
now, this workflow also runs on each pull request and acts as a gate:

- If the engineer bumps the root package but does not fulfill
  the other two requirements, then the workflow will fail and the PR
  will not be mergeable.
- If the engineer does not bump the root package but fulfills one or
  both of the other two requirements, then the workflow will not fail,
  but it will show a warning so that the engineer can double-check.
- (Pull requests that do not fulfill any of the three requirements will
  be ignored.)