fix: Strip `accessToken` from `AuthenticationController` state logs
The metadata property was recently added to the
`AuthenticationController` in #6470, but we forgot to strip out the
`accessToken`. This is currently being stripped from state logs in both
clients.

The metadata was updated to strip out this token. Additionally, some
improvements were added to the metadata snapshot tests to ensure that
all properties are represented in the fixture state.

The `@metamask/utils` package was added as a dependency because the
new metadata state deriver uses `Json` in its signature, so that type
is now used in the exported types for this package.