Allow skipping lookupNetwork call in initializeProvider
Currently, the extension patches `network-controller` so that
`initializeProvider` does not call `lookupNetwork`. Because
NetworkController gets initialized in the background process before the
UI process is initialized and because `lookupNetwork` makes a request to
the globally selected network, it was found that for a user installing
MetaMask and going through onboarding, we could share their IP address
with third parties before they had a chance to complete. That was
rightly flagged as a privacy violation, so the call to `lookupNetwork`
was disabled.

Now, long-term, `initializeProvider` shouldn't even exist, and clients
should be free to call `lookupNetwork` when they want to (and even pass
multiple network client IDs, to support a multichain flow). So at the
moment the API for NetworkController is not very ideal. However, we
don't want to keep having to patch `network-controller`.

To fix the immediate issue, then, this commit adds a `lookupNetwork`
option to `initializeProvider`. This allows clients to disable the
initial network request more straightforwardly.