Allow skipping lookupNetwork call in initializeProvider (#6575)
Currently, the extension patches `network-controller` so that
`initializeProvider` does not call `lookupNetwork`. Because
NetworkController gets initialized in the background process before the
UI process is initialized and because `lookupNetwork` makes a request to
the globally selected network, it was found that for a user installing
MetaMask and going through onboarding, we could share their IP address
with third parties before they had a chance to complete. That was
rightly flagged as a privacy violation, so the call to `lookupNetwork`
was disabled.

Now, long-term, `initializeProvider` shouldn't even exist (we can do
half of what it does in the constructor, and clients can already call
`lookupNetwork` separately). So at the moment the API for
NetworkController is not very ideal. However, we don't want to keep
having to patch `network-controller`.

For the time being, then, this commit adds a `lookupNetwork` option to
`initializeProvider`. This allows clients to disable the initial network
request more straightforwardly.