fix: Strip `accessToken` from `AuthenticationController` state logs (#6553)
## Explanation

The metadata property was recently added to the
`AuthenticationController` in #6470, but we forgot to strip out the
`accessToken`. This is currently being stripped from state logs in both
clients.

The metadata was updated to strip out this token. Additionally, some
improvements were added to the metadata snapshot tests to ensure that
all properties are represented in the fixture state.

The `@metamask/utils` package was added as a dependency because the new
metadata state deriver uses `Json` in its signature, so that type is now
used in the exported types for this package.

## References

This is an amendment to #6470

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've communicated my changes to consumers by [updating changelogs
for packages I've
changed](https://github.com/MetaMask/core/tree/main/docs/contributing.md#updating-changelogs),
highlighting breaking changes as necessary
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes