feat: Add `Messenger` error reporting
The `Messenger` class in `@metamask/messenger` has been updated to
accept an optional `captureException` constructor parameter. This is
now used to capture subscriber errors, instead of throwing them in a
`setTimeout` (which would cause a crash on mobile).

This is the implementation of the 2nd option in this ADR: https://github.com/MetaMask/decisions/blob/core-classes-error-reporting-strategy/decisions/core/0016-core-classes-error-reporting.md#optional-captureexception-constructor-parameter-that-inherits-from-parent