feat: Add `BaseController` state derivation error reporting
The state derivation functions exported by `@metamask/base-controller`
and `@metamask/base-controller/next` have been updated to accept a
`captureException` parameter. This is used to capture state derivation
errors, rather than throwing them in a `setTimeout` (which causes a
crash on mobile).

This is the implementation of the 2nd option in this ADR: https://github.com/MetaMask/decisions/blob/core-classes-error-reporting-strategy/decisions/core/0016-core-classes-error-reporting.md#optional-captureexception-constructor-parameter-that-inherits-from-parent