feat: Shared currency and token formatters (#6577)
## Explanation

This is a port of the new `Intl.NumberFormat` currency formatters from
metamask-extension. Moving it a shared library so it can be consumed
from both mobile and extension.

* What is the current state of things and why does it need to change?
There is inconsistent and multiple one-off implementations for fiat
currency and token formatting in the mobile and extension clients

* What is the solution your changes offer and how does it work?
Create a shared set of formatters that aligns mobile and extension.



## References

<!--
Are there any issues that this pull request is tied to?
https://consensyssoftware.atlassian.net/browse/CEUX-3

Are there other links that reviewers should consult to understand these
changes better?

Are there client or consumer pull requests to adopt any breaking
changes?

For example:

* Fixes #12345
* Related to #67890
-->

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've communicated my changes to consumers by [updating changelogs
for packages I've
changed](https://github.com/MetaMask/core/tree/main/docs/contributing.md#updating-changelogs),
highlighting breaking changes as necessary
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes