Expose TransactionController methods needed by SmartTransactionsController through messenger
Currently, the `confirmExternalTransaction`, `getNonceLock`,
`getTransactions`, and `updateTransaction` methods in
`TransactionController` are accessed by the
`SmartTransactionsController` constructor, but in a non-standard way:
`SmartTransactionsController` takes function references, when we would
like to call these methods via the messenger instead.

This commit exposes these methods through TransactionController's
messenger so we can do this.