feat: Add `Messenger` error reporting (#6605)
## Explanation

The `Messenger` class in `@metamask/messenger` has been updated to
accept an optional `captureException` constructor parameter. This is now
used to capture subscriber errors, instead of throwing them in a
`setTimeout` (which would cause a crash on mobile).

## References

This is the implementation of the 2nd option in this ADR:
https://github.com/MetaMask/decisions/blob/main/decisions/core/0016-core-classes-error-reporting.md#optional-captureexception-constructor-parameter-that-inherits-from-parent

Relates to #6613

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've communicated my changes to consumers by [updating changelogs
for packages I've
changed](https://github.com/MetaMask/core/tree/main/docs/contributing.md#updating-changelogs),
highlighting breaking changes as necessary
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes