fix: Ensure peer dependencies reflect latest breaking changes
Our constraints for `peerDependencies` previously only required the
major version to match. However, this was problematic for pre-1.0
packages because it allowed for situations where we cannot update a
package without introducing a peer dependency error, since pre-1.0
packages can have breaking changes in minor or patch releases.

The constraint has been updated to require the most significant part
of the version to be synchronized for peer dependencies, and all
resulting constraint errors have been auto-fixed.