fix: update start index for proposed account name for pk/hardware imports (#6677)
## Explanation

Currently, PK/Hardware imports were being started at `Account 2` as
opposed to `Account 1`. When we apply group metadata we already have
that group in state, the logic in `#applyGroupMetadata` doesn't account
for the fact that the group we're applying metadata shouldn't be counted
for in the following line:
https://github.com/MetaMask/core/blob/main/packages/account-tree-controller/src/AccountTreeController.ts#L461.
Due to the way we apply group metadata (after we've added the groups
into state), we need to take the min between `proposedNameIndex` and
`highestAccountNameIndex` NOT the max. This change in logic accounts for
group metadata being applied in the scenario when the controller is
being rehydrated through `init` and when a new pk/hardware account is
added in an active session (and handled through `accountAdded`).

## Before


https://github.com/user-attachments/assets/8ab1b3c6-574f-45bc-bf3a-3a3de7b45836

## After 


https://github.com/user-attachments/assets/66e53362-e587-45d3-b7ae-43e09da17a09

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've communicated my changes to consumers by [updating changelogs
for packages I've
changed](https://github.com/MetaMask/core/tree/main/docs/contributing.md#updating-changelogs),
highlighting breaking changes as necessary
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes