feat: decode permissions in SignatureController (#6619)
## Summary

Add execution-permission decoding to `SignatureController` for
`eth_signTypedData_v4` delegation requests. If a delegation request is
unable to be decoded into a coherent Execution Permission, reject
external (not from `origin: metamask`) attempts to sign delegations for
internal accounts. Adds decoded permission in signature request state
for rendering downstream.

## Why
Readable Permissions requires the ability to sign delegations, only when
the delegation can be decoded to a permission that is then shown to the
user.

[The architecture is described in this
document](https://www.notion.so/metamask-consensys/SignTypedData-with-Metadata-Specification-22bf86d67d688023be67e2ee06e3a56a).
Decoding has been added to `GatorPermissionsController` in #6556

Note: the request origin is always `@metamask/gator-permission-snap`
(any other origin is rejected). Because of this, `metadata` is added to
the 712 payload, defining the `justification` and `origin` which is the
_original_ origin of the request (the dapp).