feat: allow partial revokes via `wallet_revokeSession` (#6668)
## Explanation

Uniswap reported a bug in slack
[here](https://consensys.slack.com/archives/C017NBJL1S7/p1758234734587359?thread_ts=1752698865.023109&cid=C017NBJL1S7)
where MetaMask was unpermitting any EVM connections for a dapp if that
dapp used the solana wallet standard provider to make a connect request
and then cancelled it.

This call to `disconnect()` is happening outside of our packages and is
not related to a previous but similar bug where our own wallet-standard
provider was calling `wallet_revokeSession` when receiving an empty
solana `accountsChanged` event.

To fix this, this PR proposes an implementation for partial permission
revoking in the `wallet_revokeSession` handler so that we can then
update our solana `wallet-standard` provider to only revoke the solana
scopes when it is asked to disconnect.

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->

## References

<!--
Are there any issues that this pull request is tied to?
Are there other links that reviewers should consult to understand these
changes better?
Are there client or consumer pull requests to adopt any breaking
changes?

For example:

* Fixes #12345
* Related to #67890
-->

* Fixes [#728](https://consensyssoftware.atlassian.net/browse/WAPI-728)

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've communicated my changes to consumers by [updating changelogs
for packages I've
changed](https://github.com/MetaMask/core/tree/main/docs/contributing.md#updating-changelogs),
highlighting breaking changes as necessary
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes

---------

Co-authored-by: Alex Donesky <adonesky@gmail.com>