Synchronize teams.json with codeowners
The `create-update-issues` workflow uses `teams.json` to know which team
labels to assign to the tickets that it creates. Some of these team
labels were out of date with the mapping between packages and GitHub
teams in `CODEOWNERS`.

This commit corrects that. The order of packages in `teams.json` should
match the same order provided in `CODEOWNERS`, and the labels should map
to the GitHub teams.

Notably:

- `team-wallet-ux` has been split out into `team-core-extension-ux` and
  `team-mobile-ux`
- `team-wallet-framework` and `team-snaps-platform` has been combined
  into `team-core-platform`
- `team-wallet-api-platform` has been renamed to
  `team-wallet-integrations`