refactor: Remove JsonRpcEngine Result generic
Using a generic Result forced each middleware function to have the same
return type. This makes the generic useless. Supporting this, almost all
of our uses of the equivalent generic on the legacy JsonRpcMiddlware
type are Json, any, or unknown. In consequence, we remove the Result
generic.

In addition, unfreezes the result returned from asLegacyMiddleware, a
problem that was surfaced by the type refactor. Adds a test to verify
this behavior.