Fix lint:teams package script
After the Yarn upgrade, the `lint:teams` package script prints an error
indicating that the `npmMinimalAgeGate` configuration option is
unrecognized. This script uses the Yarn API to get the list of
workspaces, so it's possible something changed internally.

This commit fixes the script so that it uses `yarn workspaces` to get
information about the workspaces, hopefully avoiding any future
breakages.