chore: Update obsolete property name in error message
The `INVALID_CONTROLLER_ERROR` error message incorrectly referenced the
property name `messagingSystem`, which is no longer used in the next version of
the `BaseController` class (which the `ComposableController` switched to in
the PR `#6710).

The error message was updated to no mention the messenger instance variable at
all, since the check that triggered the error doesn't look for it anyway.