Temporarily enable `skipLibCheck` for packages
The TypeScript compiler option `skipLibCheck` has been temporarily enabled for
workspace packages. We don't want to leave this enabled forever because it can
result in dependencies silently degrading to type `any` when they have errors.
But this is necessary until we publish the next major version of the package
`base-controller`, to avoid resolving `base-controller` in dependencies to an
incompatible version.