Revert merge queue worflow changes (#6970)
## Explanation

Revert #6948 and #6949, which updated the GitHub Action workflows to
support the merge queue.

We are considering leaving the merge queue enabled, but we're reverting
this for now until we can explore it further. Specifically there was a
risk identified with our release process that we'd like to mitigate
first (where it might be possible for a release to include unreviewed
changes if it's merged with something ahead of it in the queue).

## References

Reverts #6948 and #6949

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've communicated my changes to consumers by [updating changelogs
for packages I've
changed](https://github.com/MetaMask/core/tree/main/docs/contributing.md#updating-changelogs),
highlighting breaking changes as necessary
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes