fix: Fix `wallet_revokeSession` error handling
The handler for `wallet_revokeSession` was using `instanceof` to check whether
an error matched one of two expected permission errors. This was failing in the
case where the client was using a different version of the
`@metamask/permission-controller`.

The error handling has been updated to check the error name instead of using
`instanceof`, which works correctly even if the error is sourced from a
different version of the package.