Ensure networksMetadata never references old network client IDs
In debugging a long-standing issue where SelectedNetworkController
sometimes requests invalid network client IDs from NetworkController,
one bug was discovered in NetworkController. Namely, when a network is
removed, its data is removed from `networkConfigurationsByChainId`, but
not `networksMetadata`. This commit fixes `removeNetwork` to do this,
and ensures that when the controller is initialized, if
`networksMetadata` contains invalid network client IDs, they are
removed.