ci: Fix inconsistent lint errors
Update the ESLint config for `typescript-eslint` to use the newer
`projectService` setting rather than `project`. This seems to fix the problem
we've been seeing recently with inconsistent lint results between machines, and
it's recommended by the maintainers of this plugin, and it's the default in our
shared ESLint config.

We had originally disabled this setting because we were running into OOM
errors, but these no longer occur after updating to TypeScript 5.3.