ci: Fix `claims-controller` errors in `tsconfig.json`
The `tsconfig.json` file for `claims-controller` was extending from the wrong
file, and was referencing the wrong configurations in other packages. This
caused lint errors locally in my editor, and in the PR #7078 where we are
updating how type information is read by the linter.

The issue was that the `tsconfig.build.json` files were being referenced,
which exclude test code and test files.