Fix lint:teams package script (#6941)
With the Yarn upgrade, the `lint:teams` package script now prints an
error indicating that the `npmMinimalAgeGate` configuration option is
unrecognized. This script uses the Yarn API to get the list of
workspaces, so it's possible something changed internally to where we
now need a different incantation to initialize Yarn.

This commit fixes the script so that it uses `yarn workspaces` to get
information about the workspaces instead of the Yarn API. This has
always been known to work in the past so hopefully it should avoid any
future breakages.