chore: Update ESLint to the latest version
Update ESLint to the latest version. This unblocks #6790, which will be
addressed in a later PR.

One rule adjustment was needed because the `consistent-this` rule began
flagging the use of `self` in some files written for a service worker
context. Really these shouldn't be here, but I've ignored the violation
for now and added a TODO comment to explain why.